Fixed note count for bass path of Tornado of Souls.
Added an awful hardcoded workaround to get the bass path for Tornado of Souls to return the correct note count. Currently this is the only arrangement I know of that has this bug where it reads the wrong difficulty value for a couple phrases and totally ignores the notes in them when calculating the note count. If more charts end up having this issue I will revisit this and try to find a more correct solution.